Restruct ml part
